Lockscreen charging info: show decimal in battery temperature
[neobuddy89: Adapted for Android 11]

+++ packages/SettingsLib/src/com/android/settingslib/fuelgauge/BatteryStatus.java
@@ -57,7 +57,7 @@ public class BatteryStatus {
     public final int maxChargingCurrent;
     public final int maxChargingVoltage;
     public final int maxChargingWattage;
-    public final int temperature;
+    public final float temperature;
     public final boolean present;
     public final boolean oemFastChargeStatus;
 
@@ -65,7 +65,7 @@ public class BatteryStatus {
 
     public BatteryStatus(int status, int level, int plugged, int health,
             int maxChargingWattage, int maxChargingVoltage, int maxChargingWattage,
-	    int temperature, boolean present, boolean oemFastChargeStatus, boolean dashChargeStatus) {
+	    float temperature, boolean present, boolean oemFastChargeStatus, boolean dashChargeStatus) {
         this.status = status;
         this.level = level;
         this.plugged = plugged;
